refactor: 카드 요청 검증 로직을 별도 모듈로 분리
